backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0  
sqlalchemy==2.0.25        
psycopg2-binary==2.9.9     
alembic==1.13.1          
python-dotenv==1.0.0
firebase-admin==6.2.0
pydantic==2.6.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2         